requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
Backtesting==0.3.3
bokeh==3.1.0
ccxt==3.0.45
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
dateparser==1.1.8
enum34==1.1.10
evdev==1.6.1
fonttools==4.38.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
krakenex==2.1.0
kraky
MarkupSafe==2.1.2
matplotlib==3.7.1
mido==1.2.10
mingus==0.6.1
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandas-ta==0.3.14b0
Pillow==9.4.0
PyAudio==0.2.13
pycares==4.3.0
pychord==1.2.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
pykrakenapi==0.3.1
PyMusic-Instrument==0.5
pynput==1.7.6
pyparsing==3.0.9
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.21.1
python-xlib==0.33
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==12.6.0
rtmidi==2.3.4
scikit-learn==1.2.2
scipy==1.10.1
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
synthesizer==0.2.0
ta==0.10.2
TA-Lib==0.4.25
threadpoolctl==3.1.0
tornado==6.2
typer==0.7.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.14
websocket-client==1.5.1
websockets==9.1
xyzservices==2023.2.0
yarl==1.8.2
